jocaml merged with ocaml: before-merge-luc
git-svn-id: http://caml.inria.fr/svn/ocaml/branches/jocamltrunk@6336 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ camlp4/tools/camlp4_comm.sh
@@ -3,33 +3,36 @@
 
 ARGS1=
 FILE=
+QUIET=no
 while test "" != "$1"; do
-	case $1 in
-	*.ml*) FILE=$1;;
-	*) ARGS1="$ARGS1 $1";;
-	esac
-	shift
+        case $1 in
+        -q) QUIET=yes;;
+        *.ml*) FILE=$1;;
+        *) ARGS1="$ARGS1 $1";;
+        esac
+        shift
 done
 
-head -1 $FILE >/dev/null || exit 1
+# FILE must exist and be non empty (at least one line)
+test -s "$FILE" || exit 1
 
-set - `head -1 $FILE`
+set - `awk 'NR == 1' "$FILE"`
 if test "$2" = "camlp4r" -o "$2" = "camlp4"; then
-	COMM="ocamlrun$EXE ../boot/$2$EXE -nolib -I ../boot"
+        COMM="ocamlrun$EXE ../boot/$2$EXE -nolib -I ../boot"
         if test "`basename $OTOP`" != "ocaml_stuff"; then
             COMM="$OTOP/boot/$COMM"
         fi
-	shift; shift
-	ARGS2=`echo $* | sed -e "s/[()*]//g"`
-#	ARGS1="$ARGS1 -verbose"
-	echo $COMM $ARGS2 $ARGS1 $FILE 1>&2
-	$COMM $ARGS2 $ARGS1 $FILE
+        shift; shift
+        ARGS2=`echo $* | sed -e "s/[()*]//g"`
+        ARGS1="$ARGS1 -verbose"
+        if test "$QUIET" = "no"; then echo $COMM $ARGS2 $ARGS1 $FILE; fi
+        $COMM $ARGS2 $ARGS1 $FILE
 else
-	if test "`basename $FILE .mli`.mli" = "$FILE"; then
-		OFILE=`basename $FILE .mli`.ppi
-	else
-		OFILE=`basename $FILE .ml`.ppo
-	fi
-	echo cp $FILE $OFILE 1>&2
-	cp $FILE $OFILE
+        if test "`basename $FILE .mli`.mli" = "$FILE"; then
+                OFILE=`basename $FILE .mli`.ppi
+        else
+                OFILE=`basename $FILE .ml`.ppo
+        fi
+        if test "$QUIET" = "no"; then echo cp $FILE $OFILE; fi
+        cp $FILE $OFILE
 fi
